sections/06_rag_implementation/projects/medium_analyzer/pyproject.toml
[project]
name = "medium_analyzer"
version = "0.1.0"
description = "RAG pipeline for Medium blog analysis using Pinecone vector database and Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-google-vertexai",
    "langchain-pinecone>=0.0.1",
    "pinecone-client>=3.0.0",
    "python-dotenv>=1.0.0",
    "langchainhub>=0.1.14",
    "black>=23.0.0",
    "langchain_community",
    "langchain_text_splitters",
    "langchain-classic>=1.0.0",
]

